4.3.83 Исправление ошибок PHP8 Устрожение типов Удаление неиспользуемой таблицы cms_blocks_php Организация тестов Controller_Frontend - возможность установки экземпляра диагностики Block - оптимизация вызова parent, pid; корректировка возврата значений из swap; возможность удаления второстепенных таблиц через Block::delete() CMSAccess - исправление доступа к сущностям с пустым id Material - исправление вычисления cacheFile в случае, если _SERVER['HTTPS'] задан как bool Menu - принудительная смена домена по родительскому
App.scrollTo - возможность указания id элемента
Исправление отображения placeholder в формах обратной связи у выпадающих меню

+++ tests/src/interfaces/CacheInterfaceTest.php
@@ -69,7 +69,9 @@ class CacheInterfaceTest extends BaseTest
         $testData = ['aaa' => 'bbb', 'ccc' => 'ddd', 'eee' => 'fff'];
         $interface = new CacheInterface($block, $page, [], [], [], [], [], [], $testData);
 
-        @unlink($filename);
+        if (is_file($filename)) {
+            @unlink($filename);
+        }
         $result = $interface->process();
 
         $this->assertEquals($testData, $result);
@@ -79,7 +81,9 @@ class CacheInterfaceTest extends BaseTest
 
         $this->assertEquals($testData, $cachedResult);
 
-        unlink($filename);
+        if (is_file($filename)) {
+            unlink($filename);
+        }
     }
 
 
@@ -97,7 +101,9 @@ class CacheInterfaceTest extends BaseTest
               . '<testdata>aaa</testdata>';
         $interface = new CacheInterface($block, $page, [], [], [], [], [], [], $testData);
 
-        @unlink($filename);
+        if (is_file($filename)) {
+            @unlink($filename);
+        }
         ob_start();
         echo $text;
         $result = $interface->process();
@@ -114,7 +120,9 @@ class CacheInterfaceTest extends BaseTest
             $cachedResult
         );
 
-        unlink($filename);
+        if (is_file($filename)) {
+            unlink($filename);
+        }
     }
 
 
@@ -130,7 +138,9 @@ class CacheInterfaceTest extends BaseTest
         $testData = ['aaa' => 'bbb', 'ccc' => 'ddd', 'eee' => 'fff'];
         $interface = new CacheInterface($block, $page, [], [], [], [], [], [], $testData);
 
-        @unlink($filename);
+        if (is_file($filename)) {
+            @unlink($filename);
+        }
         $result = $interface->process();
 
         $this->assertEquals($testData, $result);
